+++ src/module/vue/components/rank-hexes/rank-hexes.vue
@@ -5,7 +5,7 @@
       :key="r.rank"
       :rank="r.rank"
       :selected="r.selected"
-      @click="setRank"
+      @click="click(r.rank)"
     />
   </div>
 </template>
@@ -13,20 +13,13 @@
 <script>
 export default {
   props: {
-    actor: {
-      type: Object,
-      required: true,
-    },
-    item: {
-      type: Object,
-      required: true,
-    },
+    current: String,
   },
 
   computed: {
     ranks() {
       const keys = Object.keys(CONFIG.IRONSWORN.Ranks)
-      const position = keys.indexOf(this.item.data.rank)
+      const position = keys.indexOf(this.current)
       return keys.map((r) => {
         const rankIndex = keys.indexOf(r)
         const selected = rankIndex <= position
@@ -39,10 +32,8 @@ export default {
   },
 
   methods: {
-    setRank(_ev, rank) {
-      const actor = game.actors?.get(this.actor._id)
-      const item = actor?.items.get(this.item._id)
-      item?.update({ data: { rank } })
+    click(rank) {
+      this.$emit('click', rank)
     },
   },
 }